# Add a thread local block cache to improve read performance