--- a/README.md
+++ b/README.md
@@ -66,9 +66,9 @@ Contributions are always welcome! Here are a few tips for making a PR:
 
 ```
 cargo fmt --all -- --check
-cargo clippy --all --all-features --all-targets -- -D clippy::all
-cargo test --all --features all_except_failpoints
-cargo test --test failpoints --all-features -- --test-threads 1
+cargo +nightly clippy --all --all-features --all-targets -- -D clippy::all
+cargo +nightly test --all --features all_except_failpoints
+cargo +nightly test --test failpoints --all-features -- --test-threads 1
 ```
 
 - For changes that might induce performance effects, please quote the targeted benchmark results in the PR description. In addition to micro-benchmarks, there is a standalone [stress test tool](https://github.com/tikv/raft-engine/tree/master/stress) which you can use to demonstrate the system performance.