--- a/src/log_batch.rs
+++ b/src/log_batch.rs
@@ -855,7 +855,6 @@ mod tests {
     ) -> Vec<M::Entry> {
         let mut entries = Vec::with_capacity(entry_indexes.len());
         for ei in entry_indexes {
-            println!("{:?}", ei);
             let block =
                 LogBatch::decode_entries_block(buf, ei.entries.unwrap(), ei.compression_type)
                     .unwrap();