Micro-benchmark of fetching 100 entries from page cache:
```
Without block cache:
test engine::tests::bench_engine_fetch_entries              ... bench:     215,308 ns/iter (+/- 20,738)
test engine::tests::bench_engine_fetch_entries              ... bench:     215,171 ns/iter (+/- 25,321)

With block cache:
test engine::tests::bench_engine_fetch_entries              ... bench:      57,549 ns/iter (+/- 5,163)
test engine::tests::bench_engine_fetch_entries              ... bench:      54,762 ns/iter (+/- 4,250)
```